requirements.txt
aiogram
sqlalchemy
aiosqlite
loguru
gspread
pillow
asyncpg
fastapi
uvicorn
python-dotenv
pymorphy3
